major refactor on player, star rating, other fixes
