+++ app/src/main/java/luci/sixsixsix/powerampache2/data/local/entities/AlbumEntity.kt
@@ -33,6 +33,7 @@ fun AlbumEntity.toAlbum() = Album(
     artUrl = artUrl ?: "",
     songCount = songCount ?: 0,
     flag = flag,
+    rating = rating,
     time = time ?: 0,
     year = year ?: 0,
     genre = genre
@@ -45,6 +46,7 @@ fun Album.toAlbumEntity() = AlbumEntity(
     artists = artists,
     artistId = artist.id,
     artistName = artist.name,
+    rating = rating,
     artUrl = artUrl ?: "",
     songCount = songCount ?: 0,
     flag = flag,
