+++ app/src/main/java/luci/sixsixsix/powerampache2/data/remote/AlarmReceiver.kt
@@ -3,13 +3,11 @@ package luci.sixsixsix.powerampache2.data.remote
 import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
-import com.google.gson.Gson
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 import luci.sixsixsix.mrlog.L
-import luci.sixsixsix.powerampache2.common.toDebugString
 import luci.sixsixsix.powerampache2.domain.MusicRepository
 import luci.sixsixsix.powerampache2.domain.errors.ErrorHandler
 import javax.inject.Inject
@@ -33,6 +31,7 @@ class AlarmReceiver: BroadcastReceiver() {
         }
         // ping to refresh token
         GlobalScope.launch {
+            L("PING from AlarmReceiver")
             val ping = musicRepository.ping()
             //errorHandler.logError(Gson().toJson(ping))
             L(ping)
