+++ app/src/main/java/luci/sixsixsix/powerampache2/domain/AlbumsRepository.kt
@@ -36,4 +36,5 @@ interface AlbumsRepository {
     suspend fun getRandomAlbums(): Flow<Resource<List<Album>>>
     suspend fun getAlbum(albumId: String, fetchRemote: Boolean, ): Flow<Resource<Album>>
     suspend fun getAlbumShareLink(albumId: String): Flow<Resource<String>>
+    suspend fun rateAlbum(albumId: String, rate: Int): Flow<Resource<Any>>
 }
