+++ app/src/main/java/luci/sixsixsix/powerampache2/domain/MusicRepository.kt
@@ -22,6 +22,7 @@
 package luci.sixsixsix.powerampache2.domain
 
 import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
 import kotlinx.coroutines.flow.Flow
 import luci.sixsixsix.powerampache2.common.Resource
 import luci.sixsixsix.powerampache2.data.remote.dto.SuccessResponse
@@ -31,9 +32,10 @@ import luci.sixsixsix.powerampache2.domain.models.User
 import retrofit2.http.Query
 
 interface MusicRepository {
-
     val sessionLiveData: LiveData<Session?>
     val userLiveData: LiveData<User?>
+    val serverInfoLiveData: LiveData<ServerInfo>
+
     suspend fun ping(): Resource<Pair<ServerInfo, Session?>>
     suspend fun autoLogin(): Flow<Resource<Session>>
     suspend fun logout(): Flow<Resource<Boolean>>
