+++ app/src/main/java/luci/sixsixsix/powerampache2/presentation/screens/songs/SongsViewModel.kt
@@ -53,10 +53,12 @@ class SongsViewModel @Inject constructor(
 
             is SongsEvent.OnSongSelected -> {
                 L("SongsEvent.OnSongSelected", event.song)
-                if (playlistManager.currentQueueState.value.isEmpty()) {
-                    playlistManager.addToCurrentQueue(state.getSongList())
-                }
-                playlistManager.updateTopSong(event.song)
+                playlistManager.addToCurrentQueueUpdateTopSong(event.song, state.getSongList())
+
+//                playlistManager.updateTopSong(event.song)
+//                if (playlistManager.currentQueueState.value.isEmpty()) {
+//                    playlistManager.addToCurrentQueue(state.getSongList())
+//                }
             }
         }
     }
