+++ app/src/main/java/luci/sixsixsix/powerampache2/presentation/screens_detail/album_detail/AlbumDetailEvent.kt
@@ -32,5 +32,5 @@ sealed class AlbumDetailEvent {
     data object OnFavouriteAlbum: AlbumDetailEvent()
     data object RefreshFromCache: AlbumDetailEvent()
     data object OnShufflePlaylistToggle: AlbumDetailEvent()
-
+    data class OnNewRating(val rating: Int): AlbumDetailEvent()
 }
