+++ app/src/main/java/luci/sixsixsix/powerampache2/presentation/screens_detail/song_detail/components/SongDetailTopBar.kt
@@ -47,7 +47,7 @@ import androidx.compose.ui.unit.sp
 import luci.sixsixsix.powerampache2.R
 import luci.sixsixsix.powerampache2.common.Constants
 import luci.sixsixsix.powerampache2.presentation.common.TopBarCircularProgress
-import luci.sixsixsix.powerampache2.presentation.main.MainViewModel
+import luci.sixsixsix.powerampache2.presentation.main.viewmodel.MainViewModel
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
