+++ app/src/main/res/values/strings.xml
@@ -104,7 +104,7 @@
 
     <!---...-->
     <string name="version_quote_title">Quote of the version</string>
-    <string name="version_quote">This version is powered by black cats! debug2 delay on activity restart load song data</string>
+    <string name="version_quote">This version is powered by greens</string>
     <string name="about_appVersion_title">App Version</string>
     <string name="about_license_title">License</string>
     <string name="about_contact_title">Contact Me on Telegram and Mastodon</string>
